Rename water/reporting to water/report for consistency
