+++ message_ix_models/model/water/cli.py
@@ -276,10 +276,10 @@ def report_cli(context: "Context", output_model, sdgs, water=False):
     reg = context.model.regions
     sc = context.get_scenario()
     if water:
-        from message_ix_models.model.water.reporting import report
+        from message_ix_models.model.water.report import report
 
         report(sc, reg, sdgs)
     else:
-        from message_ix_models.model.water.reporting import report_full
+        from message_ix_models.model.water.report import report_full
 
         report_full(sc, reg, sdgs)
